improve performance of manipulating
If the make pixels functions are declared in manipulating function. The
browser needs to alloc resource for make pixels functions when we call
manipulating function, and dealloc these resource when the function
returns. If we move these make pixels function to global scope, these
functions only need to be alloc once, and all manipulating function
calls can share the same make pixels functions